Add jump distance calculation. (#1)
The reason for the magic 244 bytes was that the jump instruction is buried in the shellcode.
![image](https://user-images.githubusercontent.com/41553848/48050295-93efbf00-e1f5-11e8-932f-bbc8cafbb06a.png)

This addition calculates the required jump for any shellcode (providing there's only one jump template!). Has been tested on multiple infections. I should however note that `shellcodegen.h` is broken for me on debian 4.9.